backend/pyproject.toml
[project]
name = "salesedge-backend"
version = "1.0.0"
description = "SalesEdge - Intelligent Sales & Revenue Operations Platform"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [
    { name = "Aheli Poddar" },
    { name = "Shuvam Banerji Seal" },
    { name = "Alok Mishra" },
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.9",
    "redis[hiredis]>=5.2.0",
    "celery>=5.4.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "pandas>=2.2.0",
    "polars>=1.17.0",
    "pyarrow>=18.0.0",
    "duckdb>=1.1.0",
    "scikit-learn>=1.6.0",
    "xgboost>=2.1.0",
    "shap>=0.46.0",
    "elasticsearch>=8.17.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "alembic>=1.14.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "locust>=2.32.0",
    "playwright>=1.49.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["app"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests requiring external services",
]
